requirements.txt
# Core GUI Framework
PySide6>=6.6.0

# Image Processing
Pillow>=10.0.0

# Voice System
elevenlabs>=2.16.0
pygame>=2.6.0
python-dotenv>=1.0.0

# Platform-specific dependencies (install based on your OS)

# Windows-specific (install with: pip install pywin32)
# pywin32>=306

# macOS-specific (install with: pip install pyobjc-core pyobjc-framework-Quartz pyobjc-framework-Cocoa)
pyobjc-core>=10.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.0.0
